requirements.txt
numpy~=1.26.4
pandas~=1.4.1
scikit-learn~=1.5.1
matplotlib
gradio~=4.44.1
opencv-python
lime
shap
tqdm~=4.66.4
Pillow~=10.4.0
torch~=2.0.1
torchvision~=0.15.2
grad-cam
